requirements.txt
# Main project dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
langchain==0.0.352
langchain-openai==0.0.2
openai==1.3.8
faiss-cpu==1.12.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
requests==2.31.0
numpy==1.26.4
pandas==2.1.4
python-dotenv==1.0.0
redis==5.0.1